+++ CHANGELOG.md
@@ -1,7 +1,8 @@
-# 0.2.0 (Jan 25th 2025)
+# 0.2.0 (Jan 27th 2025)
 
 - Add environement variables support for password options
-- Add JWT support with a sign() method
+- Add JWT support with a sign() and verify() method
+- Add utility function to generate random 256-bit secrets
 
 # 0.1.0 (Dec 7th 2024)